app/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.4.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
